Rebarize erlirc.
We simply introduce the rules needed to make erlirc recognized by
rebar. While here, also clean up most of the Makefile and alter it so
it calls rebar stuff rather than call its own stuff.

The change paves the first part of the way so we can rely on the
project in other rebar'ed projects.