Clean up irc_s2c_fsm as well.
Topic is not used as a variable. While I can't make the judgemental
call if this is ok or not, I will go with Geoff's (archaelus') view
that it probably was ok.

Also comment out a function that is not currently used anywhere. We
just want it to be obvious that the function is not currently in use
in the source code.