Begin providing a more sane interface we can use at a high level.
The whole idea of #irc_cmd { ... } is good at a low level where we
want to parse and unparse messages from the irc subsystem. But at a
higher level, this is not the interface we want. Partially the problem
is that we have irc.hrl leaking the abstraction all over the place and
that we don't really want to handle all message types.

The plan here is to remedy this. We add a parser and a
format-operation which can change the underlying messages into a
high-level algebraic data type. While here, provide functions for
easily building #irc_cmd{} records. The hope is that we with time can
shield ourselves from the underlying interface via irc_cmd.erl and
thus we have the freedom to change the underlying representation at a
later time.