"Symmetrize" message sending.
We introduce irc_connection:send(pid(), irc_cmd:t()) as a way to send
irc commands in a symmetric way. Introduce an unparser for irc_cmd:t()
for the job of turning irc command ADTs into #irc_cmd{} records. While
here, limit the numbers of calls into the irc_cmd module as well to
reflect we rather want to use the ADT over the calls.